+++ lib/px/utils/pxconstants.lua
@@ -3,7 +3,7 @@
 ----------------------------------------------
 
 local _M = {
-    MODULE_VERSION = "NGINX Module v3.3.0",
+    MODULE_VERSION = "NGINX Module v4.0.0",
     RISK_PATH = "/api/v2/risk",
     CAPTCHA_PATH = "/api/v2/risk/captcha",
     ACTIVITIES_PATH = "/api/v1/collector/s2s",
@@ -13,6 +13,7 @@ local _M = {
     FIRST_PARTY_HEADER = 'x-px-first-party',
     FIRST_PARTY_VENDOR_PATH = '/init.js',
     FIRST_PARTY_XHR_PATH = '/xhr',
+    FIRST_PARTY_CAPTCHA_PATH = '/captcha',
     EMPTY_GIF_B64 = 'R0lGODlhAQABAPAAAAAAAAAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==',
 }