+++ t/009-sensitiveRoutes.t
@@ -81,7 +81,7 @@ __DATA__
         }
 
     	access_by_lua_block { 
-	    require("px.pxnginx").application()
+	    require("px.pxnginx").application(require "px.pxconfig")
 	}
 
         content_by_lua_block {
@@ -135,7 +135,7 @@ User-Agent:  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 
         }
 
     	access_by_lua_block {
-	    require("px.pxnginx").application()
+	    require("px.pxnginx").application(require "px.pxconfig")
 	}
 
         content_by_lua_block {
@@ -169,7 +169,7 @@ User-Agent:  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 
     lua_socket_pool_size 500;
     resolver 8.8.8.8;
     init_worker_by_lua_block {
-        require ("px.utils.pxtimer").application()
+        require ("px.utils.pxtimer").application(require "px.pxconfig")
     }
     set_real_ip_from   0.0.0.0/0;
     real_ip_header     X-Forwarded-For;
@@ -189,7 +189,7 @@ User-Agent:  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 
         }
 
     	access_by_lua_block {
-	    require("px.pxnginx").application()
+	    require("px.pxnginx").application(require "px.pxconfig")
 	}
 
         content_by_lua_block {
@@ -243,7 +243,7 @@ User-Agent:  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 
         }
 
     	access_by_lua_block {
-	    require("px.pxnginx").application()
+	    require("px.pxnginx").application(require "px.pxconfig")
 	}
 
         content_by_lua_block {